Code now compiles, updated such that trainingArray works with RUNNING, need to solve that if I want to add more training types

+++ trainingAnalysisTool_library/lib/trainingInstance.hpp
@@ -2,23 +2,24 @@
 #define TRAININGINSTANCE_HPP
 
 #include <string>
-#include <map>
 
 class TRAININGINSTANCE
 {
     private:
-        int m_duration;
+        double m_duration;
         std::string m_type = "";
         //Add date and time of training as well -> weekday and week number as well. For other purposes later
 
     public:
-        TRAININGINSTANCE(std::string type, int duration)
+        TRAININGINSTANCE(std::string type, double duration)
             : m_type{type}, m_duration{duration}{}
 
         std::string getType();
-        int getDuration();
+        double getDuration();
+
+        // Might be unnecessary with setters, because after object is instanciated these should not change... I think
         virtual void setType(std::string type);
-        virtual void setDuration(int duration);
+        virtual void setDuration(double duration);
 
 };
 #endif
